// warnings-service/apps/warnings-service/src/types/detailed-warning.type.ts
export type ProductType =
  | 'A'
  | 'B'
  | 'C'
  | 'D'
  | 'E'
  | 'F'
  | 'M'
  | 'O'
  | 'Q'
  | 'R'
  | 'S'
  | 'T'
  | 'W'
  | 'X';
export type ProductFullType =
  | 'Advice'
  | 'Bundle'
  | 'Climate'
  | 'Metadata'
  | 'Analysis'
  | 'Forecast'
  | 'Numerical Weather Prediction'
  | 'Observation'
  | 'Reference'
  | 'Radar'
  | 'Special'
  | 'Satellite'
  | 'Warning'
  | 'Mixed';

export type Service =
  | 'COM'
  | 'HFW'
  | 'TWS'
  | 'WAP'
  | 'WSA'
  | 'WSD'
  | 'WSF'
  | 'WSM'
  | 'WSP'
  | 'WSS'
  | 'WSW';
export type ServiceFullType =
  | 'Commercial Services'
  | 'Flood Warning Service'
  | 'Tsunami Warning Services'
  | 'Analysis and Prediction'
  | 'Aviation Weather Services'
  | 'Defence Weather Services'
  | 'Fire Weather Services'
  | 'Marine Weather Services'
  | 'Public Weather Services'
  | 'Cost Recovery Services'
  | 'Disaster Mitigation';